requirements.txt

viam-sdk==0.53.0

typing-extensions

asyncio-subprocess
